+++ changelog.md
@@ -1,3 +1,15 @@
+# release 0.0.4 (2017-11-10)
+
+- `yaml-validate`: write error message to stderr
+
+# release 0.0.3 (2017-11-10)
+
+- `yaml-validate`: executable to validate yaml-files
+
+# release 0.0.2 (2017-06-09)
+
+- Use ByteString (thanks @pbogdan)
+
 # release 0.0.1 (2016-01-06)
 
 - Initial version